minor code re-org and typo fixes
